controller, renger nad: Move code to package
Introduce pkg/render package, it includes the code for composing
NetworkAttachmentDefintion from OverlayNetwork.

No logical change has been done for the moved code.

Dockerfile is updated to copy and build pkg/.
